Fix profile mentions causing a 422 error
The Pleroma backend now reports an error when trying to reply to a status it cannot resolve we assume it is deleted or nonexistent. PleromaFE was in_reply_to_id: true as an internal method to trigger populating the post status form with the username of the profile being viewed and this was being submitted to the API as a result.


(cherry picked from commit 631c2532aafdfcbdb19ea3d48c9ce143ce7ef376)

Co-authored-by: Mark Felder <feld@feld.me>