Do not check npm version
This project does not make use of npm at all. In addition, corepack's
npm will refuse to run in a project that defines packageManager in
package.json to be yarn. If we are using standalone yarn legacy,
it will just run fine. If using corepack, it will automatically
download (if needed) and use yarn v1.